--- pkg/MasterSlave/Constants.go
@@ -1,8 +1,0 @@
-package MasterSlave
-
-var (
-	NewConnection = []byte{1}
-	// Shutdown              = []byte{2}
-	FailToConnectToTarget = []byte{3}
-	UnknownOperation      = []byte{4}
-)
